cleanup and renaming oneD_to_zeroD_convertor.py to utils.py
